requirements.txt
beautifulsoup4==4.12.3
boto3==1.36.11
botocore==1.36.11
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
duckduckgo_search==7.2.1
feedparser==6.0.11
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jmespath==1.0.1
keyring==25.6.0
keyrings.alt==5.0.2
lxml==5.3.0
more-itertools==10.6.0
-e git+https://github.com/smankoo/personal-assistant.git@51069ef3903084800d3a45028e2b58473560c778#egg=personal_assistant
primp==0.10.1
pyicloud==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.2
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
tzlocal==5.2
urllib3==2.3.0
